store content ref as string not object
i think this is the sanest breaking change to require.

the easiest way to sneak it in right now without breaking ssb-feed or
validate is to prepend to be a private message with a suffix like

.box.sha256.offchain

i can totally see and agree with an arugment to want a breaking change
on validate but this was the easierst hack to play with how this would
behave.